feat: client complains

+++ isa/src/main/java/com/example/isa/repository/BoatsRepository.java
@@ -3,5 +3,5 @@ package com.example.isa.repository;
 import com.example.isa.model.Boat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoatsRepository extends JpaRepository<Boat, Integer> {
+public interface BoatsRepository extends JpaRepository<Boat, Long> {
 }
